♻️ refactor(memory): move memorywrite/ to internal/memory/memorywrite/
Consolidate the shared profile/constraint write helpers under the memory
package. Eliminates a standalone internal/memorywrite/ package.
