+++ internal/memory/memorywrite/writer.go
