+++ internal/memory/memorywrite/writer_test.go
